requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
python-dotenv==0.21.0